Migrate GitHub example DAGs to new design #22446
