+++ docs/apache-airflow-providers-github/index.rst
@@ -39,7 +39,7 @@ Content
     :maxdepth: 1
     :caption: Resources
 
-    Example DAGs <https://github.com/apache/airflow/tree/main/airflow/providers/github/example_dags>
+    Example DAGs <https://github.com/apache/airflow/tree/main/tests/system/providers/github>
 
 .. toctree::
     :maxdepth: 1
