+++ docs/apache-airflow-providers-github/operators/index.rst
@@ -29,7 +29,7 @@ You can further process the result using **result_processor** Callable as you li
 
 An example of Listing all Repositories owned by a user, **client.get_user().get_repos()** can be implemented as following:
 
-.. exampleinclude:: /../../airflow/providers/github/example_dags/example_github.py
+.. exampleinclude:: /../../tests/system/providers/github/example_github.py
     :language: python
     :start-after: [START howto_operator_list_repos_github]
     :end-before: [END howto_operator_list_repos_github]
@@ -38,7 +38,7 @@ An example of Listing all Repositories owned by a user, **client.get_user().get_
 
 An example of Listing Tags in a Repository, **client.get_repo(full_name_or_id='apache/airflow').get_tags()** can be implemented as following:
 
-.. exampleinclude:: /../../airflow/providers/github/example_dags/example_github.py
+.. exampleinclude:: /../../tests/system/providers/github/example_github.py
     :language: python
     :start-after: [START howto_operator_list_tags_github]
     :end-before: [END howto_operator_list_tags_github]
@@ -58,14 +58,14 @@ a Tag in `GitHub <https://www.github.com/>`__.
 
 An example for tag **v1.0**:
 
-.. exampleinclude:: /../../airflow/providers/github/example_dags/example_github.py
+.. exampleinclude:: /../../tests/system/providers/github/example_github.py
     :language: python
     :start-after: [START howto_tag_sensor_github]
     :end-before: [END howto_tag_sensor_github]
 
 Similar Functionality can be achieved by directly using :class:`~airflow.providers.github.sensors.GithubSensor` ,
 
-.. exampleinclude:: /../../airflow/providers/github/example_dags/example_github.py
+.. exampleinclude:: /../../tests/system/providers/github/example_github.py
     :language: python
     :start-after: [START howto_sensor_github]
     :end-before: [END howto_sensor_github]
